the api request is now its own file
