+++ .husky/pre-commit
@@ -2,3 +2,4 @@ snyk test
 snyk code test
 npx @biomejs/biome format --write src/*.jsx
 npx prettier --write src/*.scss
+npm run build
